#1803, Change in Pickslip type int to long

+++ PickWaves/Areas/PickWaves/Controllers/ManageWavesController.cs
@@ -483,7 +483,7 @@ namespace DcmsMobile.PickWaves.Areas.PickWaves.Controllers
             }
         }
 
-        public virtual ActionResult RemovePickslipFromBucket(int bucketId, int pickslipId)
+        public virtual ActionResult RemovePickslipFromBucket(int bucketId, long pickslipId)
         {
             try
             {
